+++ src/interfaces/ICornerRadii.ts
@@ -1,0 +1,6 @@
+export default interface ICornerRadii {
+    leftTop: number
+    leftBottom: number
+    rightTop: number
+    rightBottom: number
+}