+++ src/interfaces/ISimpleColor.ts
@@ -1,0 +1,6 @@
+export default interface ISimpleColor {
+    r: number
+    g: number
+    b: number
+    a: number
+}