+++ src/interfaces/jsons/JsonDrawableText.ts
@@ -1,0 +1,8 @@
+import { TypeAndValue } from '../../utils/types'
+import IFont from '../IFont'
+import ISimpleColor from '../ISimpleColor'
+
+export default interface JsonDrawableText {
+    text?: TypeAndValue<'src' | 'text', string>
+    font?: Partial<IFont>
+}