+++ src/utils/types.ts
@@ -12,17 +12,20 @@ import ShapeableObject from '../objects/ShapeableObject'
 export type Callback = () => void
 export type SearchCallback<T> = (obj: T) => number | string
 export type EventCallback<T extends Event = Event> = (e: T) => void
-export type KeyboardEventCallback = EventCallback<KeyboardEvent>
 export type PointerCallback = (point: Point) => void
 export type Constructor<T = {}> = new (...args: any[]) => T
+export type KeyboardEventCallback = EventCallback<KeyboardEvent>
 
-// union
+// unions
 
 export type CurrentRoot = null | IRoot
+export type FollowedCameraObject = DrawablePointerable | Shape | ShapeableObject | null 
+
 export type SmoothingQuality = 'low' | 'medium' | 'high'
 export type LocalStorageEmitKeys = 'init' | 'load' | 'set'
 export type CacheDictTypes = 'image' | 'text' | 'json'
-export type FollowedCameraObject = DrawablePointerable | Shape | ShapeableObject | null 
+export type TextHorisontalAlign = 'left' | 'center' | 'right'
+export type TextVerticalAlign = 'top' | 'middle' | 'bottom'
 
 // short names
 
@@ -35,3 +38,7 @@ export type Dict<T> = Map<string, T>
 export type PointType = IPointable & ISimplePoint
 export type RectType = IRectangle & ISimpleRect
 export type DrawablePointerable = DrawableObject & PointType
+export type PositionableObject = ISimplePoint & {doWhenUpdatePosition: Function}
+
+// objects
+export type TypeAndValue<T, V> = {type: T, value: V}